move calculator to the relevant module
